# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "NomadCrew RESTful API with authentication and WebSocket support",
        "title": "NomadCrew Backend API",
        "contact": {
            "name": "NomadCrew Team",
            "url": "https://nomadcrew.uk",
            "email": "support@nomadcrew.uk"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "api.nomadcrew.uk",
    "basePath": "/v1",
    "paths": {
...
        "/invitations/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth // User must be logged in to accept": []
                    }
                ],
                "description": "Allows a user to accept an invitation to join a trip using a token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips-invitations"
                ],
                "summary": "Accept a trip invitation",
                "parameters": [
                    {
                        "description": "Invitation token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AcceptInvitationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully joined trip",
                        "schema": {
                            "$ref": "#/definitions/types.TripMembership"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid or expired token",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User ID in token does not match logged-in user (if applicable)",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found - Invitation not found or already processed",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/invitations/decline": {
            "post": {
                "security": [
                    {
                        "BearerAuth // User must be logged in to decline": []
                    }
                ],
                "description": "Allows a user to decline an invitation to join a trip using a token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips-invitations"
                ],
                "summary": "Decline a trip invitation",
                "parameters": [
                    {
                        "description": "Invitation token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.DeclineInvitationRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully declined invitation"
                    },
                    "400": {
                        "description": "Bad request - Invalid or expired token",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not authenticated or token mismatch",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found - Invitation not found or already processed",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/invitations/details": {
            "get": {
                "description": "Retrieves details about an invitation, such as the trip name and inviter, using the invitation token. Useful for UIs before acceptance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips-invitations"
                ],
                "summary": "Get details of a specific invitation using a token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invitation Token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the invitation",
                        "schema": {
                            "$ref": "#/definitions/types.InvitationDetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - Missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found - Invitation not found or token invalid",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/invitations/join": {
            "get": {
                "description": "Validates an invitation token from a deep link and redirects the user or provides invitation details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips-invitations"
                ],
                "summary": "Handle a deep link for a trip invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invitation token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirects to frontend with token or error"
                    },
                    "400": {
                        "description": "Bad request - Invalid or expired token",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found - Invitation not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/trips": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all trips that the current user is a member of",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips"
                ],
                "summary": "List user trips",
                "responses": {
                    "200": {
                        "description": "List of user's trips",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docs.TripResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new trip with the given details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips"
                ],
                "summary": "Create a new trip",
                "parameters": [
                    {
                        "description": "Trip creation details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.CreateTripRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created trip details",
                        "schema": {
                            "$ref": "#/definitions/docs.TripResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/trips/{tripId}/invitations": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invites a user by email to join a specific trip with a given role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips-invitations"
                ],
                "summary": "Invite a user to a trip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "tripId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Invitation details (email, role)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.InviteMemberRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created invitation",
                        "schema": {
                            "$ref": "#/definitions/types.TripInvitation"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid input or user already member/invited",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User does not have permission to invite",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found - Trip not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docs.ErrorResponse": {
            "description": "Error information",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "error": {
                    "description": "Detailed error information",
                    "type": "string",
                    "example": "Field 'name' is required"
                },
                "message": {
                    "description": "Error message",
                    "type": "string",
                    "example": "Invalid request parameters"
                }
            }
        },
...
        "github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.AcceptInvitationRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "handlers.DeclineInvitationRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "handlers.InviteMemberRequest": {
            "type": "object",
            "required": [
                "email",
                "role"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/types.MemberRole"
                }
            }
        },
...
        "types.TripMembership": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/types.MemberRole"
                },
                "status": {
                    "$ref": "#/definitions/types.MembershipStatus"
                },
                "tripId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}